+++ src/Cache/Implementation/WarmableCacheDecorator.php
@@ -1,0 +1,103 @@
+<?php
+
+declare(strict_types=1);
+
+/*
+ * This file is part of rekalogika/mapper package.
+ *
+ * (c) Priyadi Iman Nurcahyo <https://rekalogika.dev>
+ *
+ * For the full copyright and license information, please view the LICENSE file
+ * that was distributed with this source code.
+ */
+
+namespace Rekalogika\Mapper\Cache\Implementation;
+
+use Psr\Cache\CacheItemInterface;
+use Psr\Cache\CacheItemPoolInterface;
+use Rekalogika\Mapper\Exception\BadMethodCallException;
+use Symfony\Component\Cache\Adapter\PhpFilesAdapter;
+
+final readonly class WarmableCacheDecorator implements CacheItemPoolInterface
+{
+    private CacheItemPoolInterface $readonlyCache;
+
+    public function __construct(
+        private CacheItemPoolInterface $writableCache,
+        string $readOnlyCacheDirectory,
+        string $namespace = '',
+    ) {
+        $this->readonlyCache = new PhpFilesAdapter(
+            namespace: $namespace,
+            directory: $readOnlyCacheDirectory,
+            defaultLifetime: 0,
+            appendOnly: true,
+        );
+    }
+
+    public function getItem(string $key): CacheItemInterface
+    {
+        $item = $this->readonlyCache->getItem($key);
+
+        if (!$item->isHit()) {
+            $item = $this->writableCache->getItem($key);
+        }
+
+        return $item;
+    }
+
+    public function getWarmedUpItem(string $key): CacheItemInterface
+    {
+        return $this->readonlyCache->getItem($key);
+    }
+
+    /**
+     * @param array<array-key,string> $keys
+     * @return iterable<mixed>
+     */
+    public function getItems(array $keys = []): iterable
+    {
+        throw new BadMethodCallException('Not implemented');
+    }
+
+    public function hasItem(string $key): bool
+    {
+        return $this->readonlyCache->hasItem($key)
+            || $this->writableCache->hasItem($key);
+    }
+
+    public function clear(): bool
+    {
+        return $this->writableCache->clear();
+    }
+
+    public function deleteItem(string $key): bool
+    {
+        return $this->writableCache->deleteItem($key);
+    }
+
+    public function deleteItems(array $keys): bool
+    {
+        return $this->writableCache->deleteItems($keys);
+    }
+
+    public function save(CacheItemInterface $item): bool
+    {
+        return $this->writableCache->save($item);
+    }
+
+    public function saveWarmedUp(CacheItemInterface $item): bool
+    {
+        return $this->readonlyCache->save($item);
+    }
+
+    public function saveDeferred(CacheItemInterface $item): bool
+    {
+        return $this->writableCache->saveDeferred($item);
+    }
+
+    public function commit(): bool
+    {
+        return $this->writableCache->commit();
+    }
+}